wip: fix import cycle
