+++ pkg/registry/apis/datasource/legacy_store.go
@@ -169,7 +169,7 @@ func (s *legacyStorage) Update(ctx context.Context, name string, objInfo rest.Up
 
 		// The legacy store must use fixed names generated by the internal system
 		// we can not support external shared secrets when using the SQL backing for datasources
-		validName := v0alpha1.GetLegacySecureValueName(name, k)
+		validName := GetLegacySecureValueName(name, k)
 		if v.Name != validName {
 			return nil, false, fmt.Errorf("invalid secure value name %q, expected %q", v.Name, validName)
 		}
