+++ pkg/registry/apis/datasource/testdata/convert-dto-empty-to-resource.json
