+++ pkg/registry/apis/datasource/testdata/convert-dto-empty.json
