+++ pkg/registry/apis/datasource/testdata/convert-dto-invalid.json
