+++ pkg/registry/apis/datasource/testdata/convert-dto-testdata-to-resource.json
