+++ pkg/registry/apis/datasource/testdata/convert-resource-empty-to-cmd-add.json
