+++ pkg/registry/apis/datasource/testdata/convert-resource-empty-to-cmd-update-roundtrip.json
