+++ pkg/registry/apis/datasource/testdata/convert-resource-empty-to-legacy.json
