+++ pkg/registry/apis/datasource/testdata/convert-resource-empty.json
