+++ pkg/registry/apis/datasource/testdata/convert-resource-full-to-cmd-update.json
