+++ pkg/registry/apis/datasource/testdata/convert-resource-full-to-legacy.json
