+++ pkg/registry/apis/datasource/testdata/convert-resource-full.json
