+++ pkg/registry/apis/datasource/testdata/convert-resource-invalid.json
