+++ pkg/registry/apis/datasource/testdata/convert-resource-invalid2.json
